backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
sse-starlette==2.1.0
aiosqlite==0.20.0
